+++ src/LightBDD.Framework/Parameters/ObjectTrees/ObjectTreeNodeKind.cs
@@ -1,0 +1,7 @@
+﻿#nullable enable
+namespace LightBDD.Framework.Parameters.ObjectTrees;
+
+public enum ObjectTreeNodeKind
+{
+    Value, Array, Object
+}
