+++ src/LightBDD.Framework/Parameters/ObjectTrees/ObjectTreeValue.cs
@@ -1,0 +1,12 @@
+﻿namespace LightBDD.Framework.Parameters.ObjectTrees;
+
+public class ObjectTreeValue : ObjectTreeNode
+{
+    public ObjectTreeValue(string path, object? value) : base(path)
+    {
+        Value = value;
+    }
+
+    public override ObjectTreeNodeKind Kind => ObjectTreeNodeKind.Value;
+    public object? Value { get; }
+}
