+++ test/LightBDD.Framework.UnitTests/Parameters/ObjectTreeBuilder_tests.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using LightBDD.Framework.Parameters;
+using LightBDD.Framework.Parameters.ObjectTrees;
 using NUnit.Framework;
 using Shouldly;
